src/webhooks/grades/requirements.txt
Flask
requests
functions-framework
glom
google-cloud-firestore